Powered creeper drop
