+++ src/main/java/com/kuba6000/mobsinfo/loader/extras/Translations.java
@@ -31,7 +31,7 @@ public enum Translations {
     DRACONIC_EVOLUTION_MOB_SOUL_3,
     DRACONIC_EVOLUTION_MOB_SOUL_4,
     DRACONIC_EVOLUTION_MOB_SOUL_5,
-    EMT_CREEPER,
+    POWERED_CREEPER,
     FORBIDDEN_MAGIC_NON_PLAYER,
     OPEN_BLOCKS_SMALL_CHANCE,
     TINKERS_CONSTRUCT_BEHEADING,
