+++ src/main/resources/assets/mobsinfo/lang/en_US.lang
@@ -72,7 +72,7 @@ mobsinfo.extras.draconic_evolution_mob_soul_2=Any/None | %.2f%% | %.2f%% | %.2f%
 mobsinfo.extras.draconic_evolution_mob_soul_3=Wyvern Sword/Bow | %.2f%% | %.2f%% | %.2f%% | %.2f%% | %.2f%% | %.2f%%
 mobsinfo.extras.draconic_evolution_mob_soul_4=Draconic Sword/Bow | %.2f%% | %.2f%% | %.2f%% | %.2f%% | %.2f%% | %.2f%%
 mobsinfo.extras.draconic_evolution_mob_soul_5=Draconic Staff of Power | %.2f%% | %.2f%% | %.2f%% | %.2f%% | %.2f%% | %.2f%%
-mobsinfo.extras.emt_creeper=Drops only when Creeper is powered
+mobsinfo.extras.powered_creeper=Drops only when Creeper is powered
 mobsinfo.extras.forbidden_magic_non_player=Drops only when killed by non player entity
 mobsinfo.extras.open_blocks_small_chance=Chance: Small
 mobsinfo.extras.tinkers_construct_beheading=Each level of beheading on your weapon gives additional %.2f%%
