+++ src/main/resources/assets/mobsinfo/lang/ru_RU.lang
@@ -72,7 +72,7 @@ mobsinfo.extras.draconic_evolution_mob_soul_2=Любой/Нет           | %.2f
 mobsinfo.extras.draconic_evolution_mob_soul_3=Меч виверны/Лук     | %.2f%% |  | %.2f%% |    | %.2f%% |   | %.2f%% |    | %.2f%% |     | %.2f%% |
 mobsinfo.extras.draconic_evolution_mob_soul_4=Драконий меч/лук    | %.2f%% |  | %.2f%% |    | %.2f%% |   | %.2f%% |    | %.2f%% |     | %.2f%% |
 mobsinfo.extras.draconic_evolution_mob_soul_5=Драконий посох силы | %.2f%% |  | %.2f%% |    | %.2f%% |   | %.2f%% |    | %.2f%% |     | %.2f%% |
-mobsinfo.extras.emt_creeper=Выпадает только если крипер заряжен
+mobsinfo.extras.powered_creeper=Выпадает только если крипер заряжен
 mobsinfo.extras.forbidden_magic_non_player=Выпадает только при убийстве не игроком
 mobsinfo.extras.open_blocks_small_chance=Шанс: малый
 mobsinfo.extras.tinkers_construct_beheading=Каждый уровень обезглавливания на вашем оружии дает дополнительные %.2f%%
