+++ src/main/resources/assets/mobsinfo/lang/zh_CN.lang
@@ -53,7 +53,7 @@ mobsinfo.extras.draconic_evolution_mob_soul_2=任意/没有 | %.2f%% | %.2f%% | 
 mobsinfo.extras.draconic_evolution_mob_soul_3=双足飞龙剑/双足飞龙弓 | %.2f%% | %.2f%% | %.2f%% | %.2f%% | %.2f%% | %.2f%%
 mobsinfo.extras.draconic_evolution_mob_soul_4=神龙剑/神龙弓 | %.2f%% | %.2f%% | %.2f%% | %.2f%% | %.2f%% | %.2f%%
 mobsinfo.extras.draconic_evolution_mob_soul_5=神龙物品的能力 | %.2f%% | %.2f%% | %.2f%% | %.2f%% | %.2f%% | %.2f%%
-mobsinfo.extras.emt_creeper=当且仅当被苦力怕炸飞时，才会掉落。
+mobsinfo.extras.powered_creeper=当且仅当被苦力怕炸飞时，才会掉落。
 mobsinfo.extras.forbidden_magic_non_player=当且仅当被非玩家实体干掉时候，才会掉落。
 mobsinfo.extras.open_blocks_small_chance=几率：小
 mobsinfo.extras.tinkers_construct_beheading=每一级斩首强化都会增加%.2f%%的几率。
